feat: expose Stripe client
