+++ includes/class-stripe-connection.php
@@ -1,0 +1,76 @@
+<?php
+/**
+ * All things Stripe.
+ *
+ * @package Newspack
+ */
+
+namespace Newspack;
+
+use Stripe\Stripe;
+
+defined( 'ABSPATH' ) || exit;
+
+/**
+ * All things Stripe.
+ */
+class Stripe_Connection {
+	const STRIPE_DATA_OPTION_NAME = 'newspack_stripe_data';
+
+	/**
+	 * Get Stripe data blueprint.
+	 */
+	public static function get_default_stripe_data() {
+		return [
+			'enabled'            => false,
+			'testMode'           => false,
+			'publishableKey'     => '',
+			'secretKey'          => '',
+			'testPublishableKey' => '',
+			'testSecretKey'      => '',
+			'currency'           => 'USD',
+		];
+	}
+
+	/**
+	 * Get Stripe data, either from WC, or saved in options table.
+	 */
+	public static function get_stripe_data() {
+		$stripe_data = self::get_default_stripe_data();
+		if ( Donations::is_platform_wc() ) {
+			// If WC is configured, get Stripe data from WC.
+			$wc_configuration_manager = Configuration_Managers::configuration_manager_class_for_plugin_slug( 'woocommerce' );
+			$stripe_data              = $wc_configuration_manager->stripe_data();
+		} else {
+			$stripe_data = get_option( self::STRIPE_DATA_OPTION_NAME, self::get_default_stripe_data() );
+		}
+		$stripe_data['usedPublishableKey'] = $stripe_data['testMode'] ? $stripe_data['testPublishableKey'] : $stripe_data['publishableKey'];
+		$stripe_data['usedSecretKey']      = $stripe_data['testMode'] ? $stripe_data['testSecretKey'] : $stripe_data['secretKey'];
+		return $stripe_data;
+	}
+
+	/**
+	 * Update Stripe data. Either in WC, or in options table.
+	 *
+	 * @param object $updated_stripe_data Updated Stripe data to be saved.
+	 */
+	public static function update_stripe_data( $updated_stripe_data ) {
+		if ( Donations::is_platform_wc() ) {
+			// If WC is configured, set Stripe data in WC.
+			$wc_configuration_manager = Configuration_Managers::configuration_manager_class_for_plugin_slug( 'woocommerce' );
+			$wc_configuration_manager->update_stripe_settings( $updated_stripe_data );
+		}
+		// Otherwise, save it in options table.
+		return update_option( self::STRIPE_DATA_OPTION_NAME, $updated_stripe_data );
+	}
+
+	/**
+	 * Get Stripe client.
+	 */
+	public static function get_stripe_client() {
+		$secret_key = self::get_stripe_data()['usedSecretKey'];
+		if ( $secret_key ) {
+			return new \Stripe\StripeClient( $secret_key );
+		}
+	}
+}
